requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
sentence-transformers
scikit-learn
tf-keras
pinecone